requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pandas>=2.0.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
ollama>=0.2.0
